saucedemo/requirements.txt
selenium == 4.23.1
requests
pytest
pytest-html
python-dotenv